@@ -1,6 +1,6 @@
 module github.com/ngchain/ngcore
 
-go 1.17
+go 1.18
 
 require (
 	github.com/c0mm4nd/dbolt v0.0.0-20211212074803-fd271c7c1231
